current weather ui changed, need improvements

+++ app/src/main/java/com/onbiron/forecastmvvm/data/network/response/current/CurrentWeatherResponse.kt
@@ -43,7 +43,7 @@ data class Sys(
 data class Main(
     @SerializedName("feels_like")
     val feelsLike: Double,
-    val humidity: Int,
+    val humidity: Double,
     val pressure: Int,
     val temp: Double,
     @SerializedName("temp_max")
